Renamed to _state_checks.py for better clarity
